Add ability to pass extra params for Arxiv api
Makes the behaviour of Arxiv interactions more customisable by allowing
extra query parameters to be set in advance. This is useful when wanting
to look at the newest papers, or to include a list of specific papers
and could also be used flexibley for other configurations and to support
future query parameter options.